pyproject.toml
[tool.poetry]
name = "AI-OCR-Financial-Analyser"
version = "1.0.0"
description = ""
authors = [
    "Petr Nguyen (Main Developer) <ngn.petr@gmail.com>",
    "Samuel Kruzek (Co-Developer) <samuel.kruzek@gmail.com>",
    "Simona Dohova (Co-Developer) <s.dohova@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "1.45.1"
boto3= "1.38.27"
python-dotenv = "1.1.0"
pyyaml = "6.0.1"
botocore = "1.38.27"
pylint = "3.3.7"